+++ packages/navi-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @awell-health/navi-js-react
 
+## 0.0.5-beta.7
+
+### Patch Changes
+
+- Refactored session and auth domains; created helpers entrypoint in navi-core
+- Updated dependencies
+  - @awell-health/navi-core@0.0.5-beta.7
+  - @awell-health/navi-js@0.0.5-beta.7
+
 ## 0.0.5-beta.6
 
 ### Patch Changes