+++ packages/navi.js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @awell-health/navi-dot-js
 
+## 0.0.5-beta.7
+
+### Patch Changes
+
+- Refactored session and auth domains; created helpers entrypoint in navi-core
+- Updated dependencies
+  - @awell-health/navi-core@0.0.5-beta.7
+
 ## 0.0.5-beta.6
 
 ### Patch Changes